Stabilize generation of page and handler files with respect to blocks
